+++ source/extensions/lightspeed.trex.stage_manager.plugin.widget.usd/docs/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.2.0]
+### Added
+- Added an icon for the particle system action
+
+### Changed
+- Added the ability to create particle systems on materials
+
 ## [2.1.0]
 ## Added
 - Registering stage manager widgets in the global context menu.